Sinetone generator and audio engine hooked up to play at test tone

+++ AudioEngine/src/Common/Common.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <cstdint>
 
-typedef int32_t Volume;
+typedef float Volume;
